Split receivers test into file per test case
